requirements.txt
Django>=2.0,<3.0
djangorestframework
coreapi
pyyaml
django-filter
django-extensions
ipython
django-cors-headers
